+++ app/src/renderer/components/network/PageBlock.vue
@@ -6,11 +6,13 @@
         <div class="row">
           <div class="page-profile__header__info">
             <div class="page-profile__status-and-title">
-              <h2 class="page-profile__title">
-                Block {{ blockTitle }}
-              </h2>
+              <h2 class="page-profile__title">Block {{ blockTitle }}</h2>
               <h3 class="page-profile__subtitle">
-                {{ block.block_meta && block.block_meta.block_id ? block.block_meta.block_id.hash : '' }}
+                {{
+                  block.block_meta && block.block_meta.block_id
+                    ? block.block_meta.block_id.hash
+                    : ""
+                }}
               </h3>
             </div>
           </div>
@@ -29,7 +31,9 @@
           <div class="column">
             <dl class="info_dl colored_dl">
               <dt>Transactions</dt>
-              <dd>{{ block.block && block.block.data.txs || `No Transactions` }}</dd>
+              <dd>
+                {{ (block.block && block.block.data.txs) || `No Transactions` }}
+              </dd>
             </dl>
           </div>
         </div>
@@ -40,7 +44,12 @@
           <div class="column">
             <dl class="info_dl colored_dl">
               <dt>Evidence</dt>
-              <dd>{{ block.block && block.block.evidence.evidence || `No Evidence` }}</dd>
+              <dd>
+                {{
+                  (block.block && block.block.evidence.evidence) ||
+                    `No Evidence`
+                }}
+              </dd>
             </dl>
           </div>
         </div>
@@ -81,11 +90,11 @@ export default {
       ]
     },
     blockTitle({ num, block } = this) {
-      if (!block.block) return `n/a`
+      if (!block.block) return `--`
       return `#` + num.prettyInt(block.block.header.height)
     },
     blockTime({ moment, block } = this) {
-      if (!block.block) return `n/a`
+      if (!block.block) return `--`
       return moment(block.block.header.time).format(`MMM Do YYYY, HH:mm:ss`)
     }
   },